+++ app/src/main/java/org/sil/storyproducer/controller/export/ShareActivity.kt
@@ -14,6 +14,7 @@ import android.widget.TextView
 import com.hbisoft.pickit.Utils
 import org.sil.storyproducer.R
 import org.sil.storyproducer.controller.PopupHelpUtils
+import org.sil.storyproducer.controller.SlidePhaseFrag
 import org.sil.storyproducer.controller.phase.PhaseBaseActivity
 import org.sil.storyproducer.model.VIDEO_DIR
 import org.sil.storyproducer.model.Workspace
@@ -43,7 +44,11 @@ class ShareActivity : PhaseBaseActivity(), RefreshViewListener {
      * @return Array list of video paths
      */
     override fun onCreate(savedInstanceState: Bundle?) {
+
         super.onCreate(savedInstanceState)
+
+        SlidePhaseFrag.checkAllMarked(baseContext)
+
         setContentView(R.layout.activity_share)
         invalidateOptionsMenu()
